Add Autocomplete component and renderOption prop
