+++ packages/autocomplete/src/index.ts
@@ -1,1 +1,2 @@
 export * from './useAutocomplete'
+export * from './Autocomplete'
